NewYearCard/requirements.txt
pyqt5
pillow